refactor: move authority production leaves to daemon
