+++ packages/daemon/src/index.ts
@@ -1,5 +1,11 @@
 // TW-04 platform
 export * from "./platform/index.ts";
+export * from "./authority/production/cutover-admission.ts";
+export * from "./authority/production/production-scanner.ts";
+export * from "./authority/production/publication-evidence.ts";
+export * from "./authority/production/recovery.ts";
+export * from "./authority/production/semantic-state.ts";
+export * from "./authority/production/service-state.ts";
 export {
   ensureMachinePeopleRoster
 } from "./identity/machine-people.ts";
